requirements.txt
requests>=2.26.0
pyyaml>=5.4.1
dnspython>=2.1.0
fake-useragent>=0.1.11
urllib3==1.26.6 